Update to Java 6 (#48)
* Use enhanced for loop instead of indexed for loop
* Use enhanced for loop instead of while loop
* Remove unnecessary boxing
* Remove unnecessary unboxing

+++ src/org/isf/lab/model/LaboratoryRow.java
@@ -87,7 +87,7 @@ public class LaboratoryRow
 	        final int m = 23;
 	        int c = 133;
 	        
-	        c = m * c + (code == null ? 0 : code.intValue());
+	        c = m * c + (code == null ? 0 : code);
 	        
 	        this.hashCode = c;
 	    }
