+++ src/org/isf/patient/model/Patient.java
@@ -539,7 +539,7 @@ public class Patient {
 	        final int m = 23;
 	        int c = 133;
 	        
-	        c = m * c + ((code == null) ? 0 : code.intValue());
+	        c = m * c + ((code == null) ? 0 : code);
 	        
 	        this.hashCode = c;
 	    }
